Habilitando endpoint food and drink
