+++ src/context/GlobalProvider.js
@@ -3,15 +3,12 @@ import { useHistory } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import geral from '../data';
 import GlobalContext from './GlobalContext';
-import detailRecipes from '../data/detailRecipes';
 
 export default function GlobalProvider({ children }) {
   const [title, setTitle] = useState('');
   const [searchButton, setSearchButton] = useState(true);
   const [searchBar, setSearchBar] = useState(false);
-  const [searchTerm, setSearchTerm] = useState('');
-  const [recipesDone, setRecipesDone] = useState([]);
-  const [recipesInProgress, setRecipesInProgress] = useState([]);
+  // const [searchTerm, setSearchTerm] = useState('');
   const [state, setState] = useState(geral);
   const {
     initialState: { email, password },
@@ -24,66 +21,6 @@ export default function GlobalProvider({ children }) {
       ...prevState,
       [key]: value,
     }));
-  };
-
-  const setRecipeTitle = (value) => {
-    const newRecipeTitle = state.detailRecipes;
-    newRecipeTitle.detailsRecipe.recipeTitle = value;
-    updateState('detailRecipes', newRecipeTitle);
-  };
-
-  const setRecipeImage = (value) => {
-    const newRecipeImage = state.detailRecipes;
-    newRecipeImage.detailsRecipe.recipeImage = value;
-    updateState('detailRecipes', newRecipeImage);
-  };
-
-  const setRecipeArea = (value) => {
-    const newRecipeArea = state.detailRecipes;
-    newRecipeArea.detailsRecipe.recipeArea = value;
-    updateState('detailRecipes', newRecipeArea);
-  };
-
-  const setRecipeAlc = (value) => {
-    const newRecipeAlc = state.detailRecipes;
-    newRecipeAlc.detailsRecipe.recipeAlc = value;
-    updateState('detailRecipes', newRecipeAlc);
-  };
-
-  const setRecipeCategory = (value) => {
-    const newRecipeCategory = state.detailRecipes;
-    newRecipeCategory.detailsRecipe.recipeCategory = value;
-    updateState('detailRecipes', newRecipeCategory);
-  };
-
-  const setRecipeIngredients = (value) => {
-    const newRecipeIngredients = state.detailRecipes;
-    newRecipeIngredients.detailsRecipe.recipeIngredients = value;
-    updateState('detailRecipes', newRecipeIngredients);
-  };
-
-  const setRecipeInstructions = (value) => {
-    const newRecipeInstructions = state.detailRecipes;
-    newRecipeInstructions.detailsRecipe.recipeInstructions = value;
-    updateState('detailRecipes', newRecipeInstructions);
-  };
-
-  const setRecipeVideo = (value) => {
-    const newReciperecipeVideo = state.detailRecipes;
-    newReciperecipeVideo.detailsRecipe.recipeVideo = value;
-    updateState('detailRecipes', newReciperecipeVideo);
-  };
-
-  const setRecipeRecommendations = (value) => {
-    const newRecipeRecommendations = state.detailRecipes;
-    newRecipeRecommendations.detailsRecipe.recipeRecommendations = value;
-    updateState('detailRecipes', newRecipeRecommendations);
-  };
-
-  const setRecipeTags = (value) => {
-    const newRecipeTags = state.detailRecipes;
-    newRecipeTags.detailsRecipe.recipeTags = value;
-    updateState('detailRecipes', newRecipeTags);
   };
 
   const history = useHistory();
@@ -103,33 +40,6 @@ export default function GlobalProvider({ children }) {
         setSearchBar,
         dataFoods,
         dataDrinks,
-        getRecipeTitle: detailRecipes.detailsRecipe.recipeTitle,
-        setRecipeTitle,
-        getRecipeImage: detailRecipes.detailsRecipe.recipeImage,
-        setRecipeImage,
-        getRecipeArea: detailRecipes.detailsRecipe.recipeArea,
-        setRecipeArea,
-        getRecipeAlc: detailRecipes.detailsRecipe.recipeAlc,
-        setRecipeAlc,
-        getRecipeCategory: detailRecipes.detailsRecipe.recipeCategory,
-        setRecipeCategory,
-        getRecipeIngredients: detailRecipes.detailsRecipe.recipeIngredients,
-        setRecipeIngredients,
-        getRecipeInstructions: detailRecipes.detailsRecipe.recipeInstructions,
-        setRecipeInstructions,
-        getRecipeVideo: detailRecipes.detailsRecipe.recipeVideo,
-        setRecipeVideo,
-        getRecipeRecommendations:
-        detailRecipes.detailsRecipe.recipeRecommendations,
-        setRecipeRecommendations,
-        getRecipeTags: detailRecipes.detailsRecipe.recipeTags,
-        setRecipeTags,
-        searchTerm,
-        setSearchTerm,
-        recipesDone,
-        setRecipesDone,
-        recipesInProgress,
-        setRecipesInProgress,
         setTitle,
         setDataFoods: (value) => {
           const newInitialFoods = state.initialFoods;
